ModelsContat/requirements.txt
pytest
pytest-assume
pytest-html
pytest-timeout
pytest-repeat
allure-pytest
pyyaml
paddleclas
paddlenlp
paddlehub
h5py
attrdict
